+++ src/init.cpp
@@ -15,6 +15,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"match_font_c", (DL_FUNC) &match_font, 3},
   {"system_fonts_c", (DL_FUNC) &system_fonts, 0},
   {"dev_string_widths_c", (DL_FUNC) &dev_string_widths, 6},
+  {"dev_string_metrics_c", (DL_FUNC) &dev_string_metrics, 6},
   {NULL, NULL, 0}
 };